In the process of introducing the ability to edit played games via the API
