+++ Source/BusinessLogic/BusinessLogic.csproj
@@ -403,6 +403,7 @@
     <Compile Include="Models\Achievements\AchievementId.cs" />
     <Compile Include="Models\Games\BoardGameGeekGameCategory.cs" />
     <Compile Include="Models\Games\BoardGameGeekMechanic.cs" />
+    <Compile Include="Models\Games\UpdatedGame.cs" />
     <Compile Include="Models\PlayedGameApplicationLinkage.cs" />
     <Compile Include="Models\Games\SavePlayedGameRequest.cs" />
     <Compile Include="Models\Games\TrendingGame.cs" />
@@ -447,8 +448,8 @@
     <Compile Include="Logic\GamingGroups\IPendingGamingGroupInvitationRetriever.cs" />
     <Compile Include="Logic\GamingGroups\PendingGamingGroupInvitationRetriever.cs" />
     <Compile Include="Logic\PlayedGames\IPlayedGameDeleter.cs" />
-    <Compile Include="Logic\PlayedGames\IPlayedGameCreator.cs" />
-    <Compile Include="Logic\PlayedGames\PlayedGameCreator.cs" />
+    <Compile Include="Logic\PlayedGames\IPlayedGameSaver.cs" />
+    <Compile Include="Logic\PlayedGames\PlayedGameSaver.cs" />
     <Compile Include="Logic\PlayedGames\IPlayedGameRetriever.cs" />
     <Compile Include="Logic\PlayedGames\PlayedGameDeleter.cs" />
     <Compile Include="Logic\PlayedGames\PlayedGameRetriever.cs" />
