+++ Tests/BusinessLogic.Tests/BusinessLogic.Tests.csproj
@@ -292,6 +292,8 @@
     <Compile Include="UnitTests\LogicTests\GamingGroupsTests\GamingGroupSaverTests\UpdatePublicGamingGroupDetails.cs" />
     <Compile Include="UnitTests\LogicTests\PlayedGamesTests\ApplicationLinkerTests\ApplicationLinkerTests.cs" />
     <Compile Include="UnitTests\LogicTests\PlayedGamesTests\PlayedGameRetrieverTests\SearchPlayedGamesTests.cs" />
+    <Compile Include="UnitTests\LogicTests\PlayedGamesTests\PlayedGameSaverTests\PlayedGameSaverTestBase.cs" />
+    <Compile Include="UnitTests\LogicTests\PlayedGamesTests\PlayedGameSaverTests\UpdatePlayedGameTests.cs" />
     <Compile Include="UnitTests\LogicTests\PlayersTests\PlayerDeleterTests\DeletePlayerTests.cs" />
     <Compile Include="UnitTests\LogicTests\PlayersTests\PlayerNameBuilderTests\BuildPlayerNameTests.cs" />
     <Compile Include="UnitTests\LogicTests\PlayersTests\PlayerRetrieverTests\GetPlayerQuickStatsForUserTests.cs" />
@@ -328,7 +330,7 @@
     <Compile Include="UnitTests\LogicTests\GamingGroupsTests\GamingGroupRetrieverTests\GetGamingGroupDetailsTests.cs" />
     <Compile Include="UnitTests\LogicTests\GamingGroupsTests\GamingGroupSaverTests\GamingGroupSaverTestBase.cs" />
     <Compile Include="UnitTests\LogicTests\GamingGroupsTests\PendingGamingGroupInvitationRetrieverImplTests\GetPendingGamingGroupInvitationsTests.cs" />
-    <Compile Include="UnitTests\LogicTests\PlayedGamesTests\PlayedGameCreatorTests\CreatedPlayedGameTests.cs" />
+    <Compile Include="UnitTests\LogicTests\PlayedGamesTests\PlayedGameSaverTests\CreatedPlayedGameTests.cs" />
     <Compile Include="IntegrationTests\LogicTests\Players\GetTopPlayersIntegrationTestsTests.cs" />
     <Compile Include="UnitTests\LogicTests\PlayedGamesTests\PlayedGameDeleterTests\DeletePlayedGameTests.cs" />
     <Compile Include="UnitTests\LogicTests\NemesesTests\NemesisRecalculatorTests\RecalculateNemesisTests.cs" />
